# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Public Transit Management System REST API",
    "description": "Registro, inicio de sesión, CRUD estaciones, CRUD rutas"
  },
  "host": "https://public-transit.onrender.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer &lt;token>**"
    }
  },
  "paths": {
    "/api/users/": {
      "post": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "address2": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                },
                "state": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "zip": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "description": "",
        "tags": ["User"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/login/": {
      "post": {
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/stations/": {
      "post": {
        "description": "",
        "tags": ["Station"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "latitude": {
                  "example": "any"
                },
                "longitude": {
                  "example": "any"
                },
                "capacity": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "description": "",
        "tags": ["Station"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/stations/{id}": {
      "get": {
        "description": "",
        "tags": ["Station"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Station"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Station"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "latitude": {
                  "example": "any"
                },
                "longitude": {
                  "example": "any"
                },
                "capacity": {
                  "example": "any"
                },
                "estado": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/routes/": {
      "post": {
        "description": "",
        "tags": ["Route"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "distance": {
                  "example": "any"
                },
                "stations": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "description": "",
        "tags": ["Route"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/routes/{id}": {
      "get": {
        "description": "",
        "tags": ["Route"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Route"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Route"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "distance": {
                  "example": "any"
                },
                "stations": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vehicles/{type}": {
      "post": {
        "description": "",
        "tags": ["Vehicle"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          }, 
          {  
            "name": "body",
            "in": "body",
            "schema" : {
              "type": "object",
              "properties" : {
                "capacity": {
                  "example": "Number"
                },
              "fuel": {
                "example": "Number"
              },
              "plate": {
                "example": "String"
              },
              "model": {
                "example": "String"
              },
              "maintenances": {
                "example" : "[]"
              } 
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "description": "",
        "tags": ["Vehicle"],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vehicles/": {
      "get": {
        "description": "",
        "tags": ["Vehicle"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/vehicles/{id}": {
      "get": {
        "description": "",
        "tags": ["Vehicle"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Vehicle"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/vehicles/{type}/{id}": {
      "put": {
        "description": "",
        "tags": ["Vehicle"],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {  
            "name": "body",
            "in": "body",
            "schema" : {
              "type": "object",
              "properties" : {
                "capacity": {
                  "example": "Number"
                },
              "fuel": {
                "example": "Number"
              },
              "plate": {
                "example": "String"
              },
              "model": {
                "example": "String"
              },
              "maintenances": {
                "example" : "[]"
              } 
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenances/": {
      "post": {
        "description": "",
        "tags": ["Maintenance"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "example": "any"
                },
                "cost": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "description": "",
        "tags": ["Maintenance"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/maintenances/{id}": {
      "get": {
        "description": "",
        "tags": ["Maintenance"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Maintenance"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Maintenance"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "example": "any"
                },
                "cost": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/schedules/": {
      "post": {
        "description": "",
        "tags":["Schedule"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "departureTimes": {
                  "example": "any"
                },
                "route": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "description": "",
        "tags":["Schedule"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/schedules/{id}": {
      "get": {
        "description": "",
        "tags":["Schedule"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "tags":["Schedule"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "",
        "tags":["Schedule"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "departureTimes": {
                  "example": "any"
                },
                "route": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}